requirements.txt
mwparserfromhell>=0.6.6
elasticsearch>=8.12.0
psycopg2-binary>=2.9.9
anthropic>=0.20.0
fastapi>=0.110.0
uvicorn>=0.29.0
tqdm>=4.66.0
python-dotenv>=1.0.0
slowapi>=0.1.9
pytest>=8.0.0
sqlmodel>=0.0.14
PyJWT>=2.8.1
authlib>=1.3.0
python-multipart>=0.0.6
